requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
PyJWT==2.9.0
resend==2.4.0
python-multipart==0.0.9      # Required for UploadFile support in FastAPI
